@@ -4,6 +4,7 @@ import org.springframework.boot.diagnostics.AbstractFailureAnalyzer;
 import org.springframework.boot.diagnostics.FailureAnalysis;
 
 import com.giffing.bucket4j.spring.boot.starter.exception.Bucket4jGeneralException;
+import com.giffing.bucket4j.spring.boot.starter.exception.ExecutePredicateBeanNotFoundException;
 import com.giffing.bucket4j.spring.boot.starter.exception.FilterURLInvalidException;
 import com.giffing.bucket4j.spring.boot.starter.exception.JCacheNotFoundException;
 import com.giffing.bucket4j.spring.boot.starter.exception.MissingKeyFilterExpressionException;
@@ -34,11 +35,15 @@ public class Bucket4JAutoConfigFailureAnalyzer extends AbstractFailureAnalyzer<B
 			actionMessage = "Please set the property 'expression' in your configuration file with a valid expression (see Spring Expression Language)" + NEW_LINE;
 		}
 		
-		if( cause instanceof FilterURLInvalidException) {
-			FilterURLInvalidException e = (FilterURLInvalidException) cause;
+		if( cause instanceof FilterURLInvalidException e) {
 			descriptionMessage = "You've set an invalid regular expression in the 'filter'" + NEW_LINE +
 					"Description: " + e.getDescription();
 			actionMessage = "To Filter for all requests type .* or remove the property 'bucket4j.filters.url' completly." + NEW_LINE;
+		}
+		
+		if( cause instanceof ExecutePredicateBeanNotFoundException e) {
+			descriptionMessage = "You've configured the '%s=' execution predicate which doesn't exists.".formatted(e.getExecutePredicateName());
+			actionMessage = "Please check if you've spelled it correctly or provide a coressponding ExecutePredicate bean." + NEW_LINE;
 		}
 		
 		return new FailureAnalysis(descriptionMessage, actionMessage, cause);
